Rocket file: Dispatch the toggleRocket with id and moving files to rockets folder
