+++ src/components/rockets/Rockets.js
@@ -2,13 +2,13 @@ import React, { useEffect } from 'react';
 import { useSelector, shallowEqual, useDispatch } from 'react-redux';
 import Rocket from './Rocket';
 import styles from './Rockets.module.scss';
-import { loadRockets } from '../redux/rockets/rockets';
+import { loadRockets } from '../../redux/rockets/rockets';
 
 const Rockets = () => {
   const dispatch = useDispatch();
   const rockets = useSelector((state) => state.rockets.data, shallowEqual);
   const loading = useSelector((state) => state.rockets.isLoading, shallowEqual);
-  useEffect(() => { dispatch(loadRockets()); }, []);
+  useEffect(() => rockets.length === 0 && dispatch(loadRockets()), []);
 
   const loadingMessage = <div> Loading rockets...</div>;
   const content = (rockets) => (
